divided navbar into small & large. Added scale(0.05) + transition to .card:hover
