+++ components/card.module.scss
@@ -33,5 +33,6 @@
   }
   &:hover {
     cursor: pointer;
+    transform: scale(1.05);
   }
 }
