+++ components/navbar.js
@@ -1,6 +1,6 @@
-import styles from "./navbar.module.scss";
 import { useState, useCallback, useEffect, useRef } from "react";
-import { useRouter } from "next/router";
+import NavbarSmall from "./navbarSmall";
+import NavbarLarge from "./navbarLarge";
 
 const Navbar = () => {
   //För att läsa av skärmstorlek och sätta isBreakpoint till true/false. (isBreakpoint -> rad. 32)
@@ -30,106 +30,18 @@ const Navbar = () => {
   };
 
   const isBreakpoint = useMediaQuery(768);
-  const [dropdownVisible, setDropdownVisible] = useState(false);
-  const dropdown = useRef(null);
 
-  //För att toggla dropdownmenu när vi klickar på hamburger Menu.
-  const handleonClick = () => {
-    setDropdownVisible(!dropdownVisible);
-  };
-
-  //För att stänga dropdownmenu när den är öppen & vi klickar utanför den.
-  useEffect(() => {
-    if (!dropdownVisible) return;
-
-    function handleClick(event) {
-      if (dropdown.current && !dropdown.current.contains(event.target)) {
-        setDropdownVisible(false);
-      }
-    }
-    const timeoutId = setTimeout(() => {
-      window.addEventListener("click", handleClick);
-    }, 0);
-
-    return () => {
-      clearTimeout(timeoutId);
-      window.removeEventListener("click", handleClick);
-    };
-  }, [dropdownVisible]);
-
-  //För att få aktiva Länkar.
-  let aboutActive = false;
-  let homeActive = false;
-  const router = useRouter();
-  if (router.pathname == "/about") {
-    aboutActive = true;
-  } else if (router.pathname == "/") {
-    homeActive = true;
-  }
-
-  return (
-    <>
-      {isBreakpoint ? (
-        <>
-          <header className={styles.navbar}>
-            <img
-              className={styles.hamburgerMenu}
-              src="/HamburgerMenu.svg"
-              alt="hamburger menu"
-              width="69px"
-              height="69px"
-              onClick={handleonClick}
-            />
-          </header>
-          <div
-            className={styles.dropdown}
-            style={{ display: dropdownVisible ? "block" : "none" }}
-            ref={dropdown}
-          >
-            <a
-              className={homeActive ? styles.linkActive : styles.link}
-              href="./"
-            >
-              <h2>Home</h2>
-            </a>
-            <h2>Group</h2>
-            <h2>Shopping List</h2>
-            <a
-              className={aboutActive ? styles.linkActive : styles.link}
-              href="./about"
-            >
-              <h2>About Us</h2>
-            </a>
-          </div>
-        </>
-      ) : (
-        <header className={styles.navbar}>
-          <a className={homeActive ? styles.linkActive : styles.link} href="./">
-            <h2>Home</h2>
-          </a>
-          <h2>Group</h2>
-          <h2>Shopping List</h2>
-          <a
-            className={aboutActive ? styles.linkActive : styles.link}
-            href="./about"
-          >
-            <h2>About Us</h2>
-          </a>
-        </header>
-      )}
-    </>
-  );
+  return <>{isBreakpoint ? <NavbarSmall /> : <NavbarLarge />}</>;
 };
 
 export default Navbar;
 
 // import styles from "./navbar.module.scss";
-// import { useState, useCallback, useEffect } from "react";
-// // import { NavLink } from "next";
-// //Stand- up
-// //Behövs detta för att få aktiva länkar?
+// import { useState, useCallback, useEffect, useRef } from "react";
+// import { useRouter } from "next/router";
 
 // const Navbar = () => {
+//   //För att läsa av skärmstorlek och sätta isBreakpoint till true/false. (isBreakpoint -> rad. 32)
 //   const useMediaQuery = (width) => {
 //     const [targetReached, setTargetReached] = useState(false);
 
@@ -157,12 +69,41 @@ export default Navbar;
 
 //   const isBreakpoint = useMediaQuery(768);
 //   const [dropdownVisible, setDropdownVisible] = useState(false);
+//   const dropdown = useRef(null);
 
+//   //För att toggla dropdownmenu när vi klickar på hamburger Menu.
 //   const handleonClick = () => {
 //     setDropdownVisible(!dropdownVisible);
 //   };
-//   //Stand- Up
-//   //Vill ha bort menyn när vi klickar utanför.
+
+//   //För att stänga dropdownmenu när den är öppen & vi klickar utanför den.
+//   useEffect(() => {
+//     if (!dropdownVisible) return;
+
+//     function handleClick(event) {
+//       if (dropdown.current && !dropdown.current.contains(event.target)) {
+//         setDropdownVisible(false);
+//       }
+//     }
+//     const timeoutId = setTimeout(() => {
+//       window.addEventListener("click", handleClick);
+//     }, 0);
+
+//     return () => {
+//       clearTimeout(timeoutId);
+//       window.removeEventListener("click", handleClick);
+//     };
+//   }, [dropdownVisible]);
+
+//   //För att få aktiva Länkar.
+//   let aboutActive = false;
+//   let homeActive = false;
+//   const router = useRouter();
+//   if (router.pathname == "/about") {
+//     aboutActive = true;
+//   } else if (router.pathname == "/") {
+//     homeActive = true;
+//   }
 
 //   return (
 //     <>
@@ -181,25 +122,35 @@ export default Navbar;
 //           <div
 //             className={styles.dropdown}
 //             style={{ display: dropdownVisible ? "block" : "none" }}
+//             ref={dropdown}
 //           >
-//             <a className={styles.link} href="./">
+//             <a
+//               className={homeActive ? styles.linkActive : styles.link}
+//               href="./"
+//             >
 //               <h2>Home</h2>
 //             </a>
 //             <h2>Group</h2>
 //             <h2>Shopping List</h2>
-//             <a className={styles.link} href="./about">
+//             <a
+//               className={aboutActive ? styles.linkActive : styles.link}
+//               href="./about"
+//             >
 //               <h2>About Us</h2>
 //             </a>
 //           </div>
 //         </>
 //       ) : (
 //         <header className={styles.navbar}>
-//           <a className={styles.link} href="./">
+//           <a className={homeActive ? styles.linkActive : styles.link} href="./">
 //             <h2>Home</h2>
 //           </a>
 //           <h2>Group</h2>
 //           <h2>Shopping List</h2>
-//           <a className={styles.link} href="./about">
+//           <a
+//             className={aboutActive ? styles.linkActive : styles.link}
+//             href="./about"
+//           >
 //             <h2>About Us</h2>
 //           </a>
 //         </header>
