+++ components/navbar.module.scss
@@ -10,7 +10,6 @@
 }
 
 .link {
-  display: block;
   text-decoration: none;
   color: rgb(255, 255, 255);
   transition: all 0.2s ease-in-out;
